test(infrastructure): refactor class
